pyproject.toml
[project]
name = "openclaw-cert-cli"
version = "2.0.0"
description = "OpenClaw Master-Zertifizierungs-CLI — Modulare Paketstruktur"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
]

[project.scripts]
openclaw-cert = "openclaw_cert.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=openclaw_cert",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "slow: langsame Tests (z.B. Thread-Tests)",
]

[tool.coverage.run]
source = ["openclaw_cert"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "if __name__",
    "pragma: no cover",
]
